Remove FinalFormLayoutDisplay and move logic to FinalFormLayoutSelect
